backened/pyproject.toml
[tool.poetry]
name = "todo-auth-react-dj"
version = "0.1.0"
description = ""
authors = ["Alisher007 <alisherx07@gmail.com>"]
readme = "README.md"
packages = [{include = "todo_auth_react_dj"}]

[tool.poetry.dependencies]
python = "^3.10"
asgiref = "3.7.2"
django = "4.2.5"
django-cors-headers = "4.2.0"
djangorestframework = "3.14.0"
pytz = "2023.3.post1"
sqlparse = "0.4.4"
typing-extensions = "4.7.1"
djangorestframework-simplejwt = "^5.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
